+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ### 1.0.5 patch
 * Add Ukrainian and Esperanto language packs (still to be proof-read)
-* Keyboard shortcut (`CTRL + ALT + SPACE`) to switch focus between notation and encoding panel (#55)
+* Keyboard shortcut (`SHIFT + SPACE`) to switch focus between notation and encoding panel (#55)
 * Keyboard shortcut (`N`) to duplicate a selected chord, note or rest, or insert a new note into an empty layer (#57)
 * Keyboard shortcut (`SHIFT + N`) to convert selected notes to rests and rests to notes (#57)
 * Keyboard shortcut (`.`) to add or remove `dots="1"` to notes, chords, etc.
 * Keyboard shortcut (`C`) to embed selected notes inside a new chord element, and vice versa (remove chord and leave notes)
+* Misleading warning when clicking on elements without ID fixed (#52)
 
 ### 1.0.4 patch
 * Tweak CSS select and translations (German)